feat: enhance authentication flow and refactor user management
- Added `apiUrl` to configuration for dynamic API endpoint usage.
- Refactored import paths for authentication middleware and user utilities to improve structure and maintainability.
- Introduced new `Login` component for user authentication with magic link and social sign-in options.
- Implemented `getUser` utility to streamline user retrieval across the application.
- Updated API routes to utilize the new user management logic, ensuring consistent user access checks.
- Created new files for authentication middleware and Supabase client handling, enhancing the overall architecture of the authentication system.
