+++ app/(login)/sign-in/page.tsx
@@ -1,6 +1,6 @@
-import { getUser } from '@/queries/user';
+import { getUser } from '@/utils/user';
 import { redirect } from 'next/navigation';
-import Login from '../login';
+import Login from '../../../features/auth/components/login';
 
 const SignIn = async () => {
   const user = await getUser();
