+++ app/(login)/sign-up/page.tsx
@@ -1,6 +1,6 @@
-import { getUser } from '@/queries/user';
+import { getUser } from '@/utils/user';
 import { redirect } from 'next/navigation';
-import Login from '../login';
+import Login from '../../../features/auth/components/login';
 
 const SignUp = async () => {
   const user = await getUser();
@@ -8,6 +8,7 @@ const SignUp = async () => {
   if (user) {
     return redirect('/dashboard');
   }
+
   return <Login mode="signup" />;
 };
