+++ app/api/book/[id]/route.ts
@@ -1,7 +1,7 @@
 import { db } from '@/db';
 import { BooksTable } from '@/db/schema';
 import { APIResponse } from '@/types/api';
-import { getUserByClerkId } from '@/utils/auth';
+import { getUser } from '@/utils/user';
 import { and, eq } from 'drizzle-orm';
 import { NextResponse } from 'next/server';
 
@@ -9,12 +9,13 @@ export const DELETE = async (
   req: Request,
   { params }: { params: { id: string } }
 ) => {
-  const { id } = await getUserByClerkId();
+  const user = await getUser();
+  if (!user) throw new Error('User not found');
 
   try {
     await db
       .delete(BooksTable)
-      .where(and(eq(BooksTable.id, params.id), eq(BooksTable.userId, id)));
+      .where(and(eq(BooksTable.id, params.id), eq(BooksTable.userId, user.id)));
   } catch (error) {
     console.error('Failed to delete book', error);
     return NextResponse.json<APIResponse<null>>(
