+++ config.ts
@@ -1,6 +1,7 @@
 const config = {
   appName: 'Kindian',
   domainName: 'http://localhost:3000',
+  apiUrl: 'http://localhost:4000',
 };
 
 export default config;
