+++ features/auth/type.ts
@@ -1,0 +1,1 @@
+export type AuthMode = 'signin' | 'signup';
