+++ features/auth/utils/constants.ts
@@ -1,0 +1,3 @@
+import config from '@/config';
+
+export const REDIRECT_TO_AUTH_CALLBACK = `${config.domainName}/api/auth/callback`;
